pyproject.toml
[tool.poetry]
name = "procedural-pixel-art-kata"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^7.2.0"
requests = "^2.24.0"
lxml = "^4.5.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
mypy = "^0.782"
pytest = "^6.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
